refactor(components/molecule/validationCode): api tab
