+++ components/molecule/validationCode/demo/articles/ArticleDefault.js
@@ -1,7 +1,7 @@
 import PropTypes from 'prop-types'
 import {Article, H2, Paragraph, Input} from '@s-ui/documentation-library'
 import {useState} from 'react'
-import MoleculeValidationCode from '../../src/validationCode'
+import MoleculeValidationCode from '../../src'
 
 const ArticleDefault = ({className}) => {
   const [code, setCode] = useState([])
