+++ components/molecule/validationCode/demo/articles/ArticleLength.js
@@ -8,7 +8,7 @@ import {
   Cell
 } from '@s-ui/documentation-library'
 import {useState} from 'react'
-import MoleculeValidationCode from '../../src/validationCode'
+import MoleculeValidationCode from '../../src'
 
 const ArticleLenght = ({className}) => {
   const [length, setLength] = useState(6)
