+++ components/molecule/validationCode/demo/articles/ArticlePlaceholder.js
@@ -9,7 +9,7 @@ import {
   Label
 } from '@s-ui/documentation-library'
 import {useState} from 'react'
-import MoleculeValidationCode from '../../src/validationCode'
+import MoleculeValidationCode from '../../src'
 
 const ArticlePlaceHolder = ({className}) => {
   const [code, setCode] = useState('1234')
