+++ components/molecule/validationCode/demo/articles/ArticleUncontrolled.js
@@ -9,7 +9,7 @@ import {
   Paragraph
 } from '@s-ui/documentation-library'
 import {useState} from 'react'
-import MoleculeValidationCode from '../../src/validationCode'
+import MoleculeValidationCode from '../../src'
 
 const cellStyles = {
   display: 'flex',
