+++ components/molecule/validationCode/src/config.js
@@ -3,3 +3,12 @@ export {
   pinInputSizes as validationCodeSizes,
   pinInputMask as validationCodeMask
 } from '@s-ui/react-atom-pin-input'
+
+export const BASE_CLASS = 'sui-MoleculeValidationCode'
+
+export const normalizeValue = value => {
+  if (value === undefined) {
+    return value
+  }
+  return `${typeof value === 'string' ? value.split('') : value}`
+}
